Smarter directories

+++ scripts/copysite.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
+# Get repo root
+repo_root=$(git rev-parse --show-toplevel 2>/dev/null)
+if [ -z "$repo_root" ]; then
+    echo "Not in a Git repository."
+    exit
+fi
+
 sudo rm -fr /var/www/html/wspr
 sudo mkdir /var/www/html/wspr
-sudo cp -R /home/pi/WsprryPi/data/* /var/www/html/wspr/
+sudo cp -R "$repo_root"/WsprryPi/data/* /var/www/html/wspr/
 sudo ln -sf /usr/local/etc/wspr.ini /var/www/html/wspr/wspr.ini
 sudo chown -R www-data:www-data /var/www/html/wspr/
